requirements.txt
numpy
fire
transformers
torch
sentencepiece
tokenizers>=0.13.3
GitPython
